refactor: extract pure rendering engine into render.ts; drop dead tslint.json
The number→words logic doesn't depend on instance state, so it no longer
lives on the Tafgeet class. Moved the pure functions (readNumber,
renderIntegerWords, renderCountedNoun, and their helpers) into a dedicated
src/render.ts module; the class now owns only input parsing/validation and the
public parse()/read() surface. Behavior is byte-identical — 1064 tests
unchanged.

Also removed tslint.json, dead since the v1.1 migration to ESLint 9, and
refreshed two stale doc comments in types.ts (gender is intrinsic, not an
option; scale-word plural covers 3–10).
