--- tslint.json
@@ -1,4 +1,0 @@
-{
-  "extends": ["tslint:recommended", "tslint-config-prettier"],
-  "only-arrow-functions": false
-}
